Rename useBookmarks to useFetchBookmarks
