+++ packages/ui/src/components/domain/sidebar/NavSidebar.tsx
@@ -3,7 +3,7 @@ import { LuLogOut } from 'react-icons/lu'
 import { withRouter } from 'next/router'
 import Sidebar from '@/components/domain/sidebar'
 import Skeleton from '@/components/shared/Skeleton'
-import useBookmarks from '@/hooks/useBookmarks'
+import useBookmarks from '@/hooks/useFetchBookmarks'
 import Group from './Group'
 import Item from './Item'
 import MobileMenu from './MobileMenu'
